requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.1
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.3
MarkupSafe==2.1.2
Pillow==9.5.0
PyJWT==2.7.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.0.2
